Inject API via manifest content script to ensure it is reliably executed before page
Previous (uncommited) attempts via the scripting API did work, but not
reliably. In some instances, the injected script (from the onCommited
webNavigation handler in the background script) was injected after the
page's script was executed. Also, the now used approach has the
advantage that it does not require scripting and webNavigation
permissions.