WIP: (missing ciphertext validation!) hash of concatenated sign+enc public key for recipient key?
This approach has the advantage that it does not require the signature
in the recipient key information. On the other hand, this means that it
is required to have both the encryption and signing key pair of the
respective party in order to verify the key id. This is not the case if
a user does not have a specific recipient key information of some
sender. Thus, it would become necessary to include the encryption public
key in the ciphertext data for every single message as well. As I assume
there are generally much more message exchanges than recipient key
information exchanges, this would be the worse trade-off.